refactor: /my/favorite 페이지 전환
